+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Added "Curse" effect; temporarily preventing the streamer from using any items.
+- Changed trap effects to retry if the player is already in the trap.
 
 ## [1.0.9] - 2022-12-11